+++ fe/src/components/organisms/walk/WalkDetailModal.tsx
@@ -7,8 +7,8 @@ import Button from '@/components/atoms/Button';
 import WalkDetailFormItem from '@/components/molecules/walk/WalkDetailFormItem';
 import { NAV_HEIGHT } from '@/components/organisms/Nav';
 import { FormDataPatchType } from '@/components/organisms/walk/WalkModal';
-import { useWalkOnSuccess } from '@/hooks/useWalkOnSuccess';
-import { useWalkDeleteMutation, useWalkPatchMutation } from '@/hooks/useWalkQuery';
+import { useWalkOnSuccess } from '@/hooks/walk/useWalkOnSuccess';
+import { useWalkDeleteMutation, useWalkPatchMutation } from '@/hooks/walk/useWalkQuery';
 import { usePetStore } from '@/stores/usePetStore';
 import { theme } from '@/styles/theme';
 import TrashIcon from '@/svg/trash.svg';
