+++ fe/src/components/organisms/walk/WalkModal.tsx
@@ -7,8 +7,8 @@ import styled from 'styled-components';
 import Button from '@/components/atoms/Button';
 import WalkFormItem from '@/components/molecules/walk/WalkFormItem';
 import { NAV_HEIGHT } from '@/components/organisms/Nav';
-import { useWalkOnSuccess } from '@/hooks/useWalkOnSuccess';
-import { useWalkMutation } from '@/hooks/useWalkQuery';
+import { useWalkOnSuccess } from '@/hooks/walk/useWalkOnSuccess';
+import { useWalkMutation } from '@/hooks/walk/useWalkQuery';
 import { theme } from '@/styles/theme';
 import TrashIcon from '@/svg/trash.svg';
 import { BuddysType, PositionType, SelectedBuddysType, TimeRef } from '@/types/map';
