+++ fe/src/components/pages/walk/Walk.tsx
@@ -11,7 +11,7 @@ import { NAV_HEIGHT } from '@/components/organisms/Nav';
 import WalkCalendar from '@/components/organisms/walk/WalkCalendar';
 import WalkDetailModal from '@/components/organisms/walk/WalkDetailModal';
 import WalkList from '@/components/organisms/walk/WalkList';
-import { useWalkQuery, UseWalkQueryProps } from '@/hooks/useWalkQuery';
+import { useWalkQuery, UseWalkQueryProps } from '@/hooks/walk/useWalkQuery';
 // import { testData } from '@/mocks/walkTest';
 import { useFilterStore } from '@/stores/useFilterStore';
 import { usePetStore } from '@/stores/usePetStore';
